Section_098_-_Day_98_-_Professional_Portfolio_Project_-_[Python_Automation]/1_raise_email_automation/requirements.txt
schedule==1.2.1
python-dotenv==1.0.1
APScheduler==3.10.4